Refactor engineering functions and pipeline
